💫💝 Checkpoint ./bot.js:26490528/2

+++ bot.js
@@ -323,7 +323,7 @@ jaillog.send(embed)
             }
       const reklam2 = [".net",".com",".tk","www.","https://","http://",".png",".gif",".webp"];
     if (reklam2.some(word => message.content.toLowerCase().includes(word))) {
-    //if (message.member.roles.cache.get("976890819518820372")) return
+    if (message.member.roles.cache.get("976890819518820372")) return
     if (message.author.bot) return
         message.delete();
         message.channel.send(new Discord.MessageEmbed().setDescription(`${message.author}, Sunucumuzda link paylaşmak yasak.`).setFooter(`Atahan`).setAuthor(message.guild.name, message.guild.iconURL({dynamic: true})).setColor('GRAY')).then(x => x.delete({timeout: 5000}))
